+++ lib/libfaeris/src/support/util/FsPathUtil.cc
@@ -45,7 +45,7 @@ bool PathUtil::hasExtend(const char* name)
 std::string PathUtil::getDirName(const char* name)
 {
 	int split=0;
-	int name_len=strlen(name);
+	int name_len=(int)strlen(name);
 	for(int i=name_len-1;i>=0;i--)
 	{
 		if(name[i]=='/')
@@ -60,7 +60,7 @@ std::string PathUtil::getDirName(const char* name)
 std::string PathUtil::getFileName(const char* name)
 {
 	int split=-1;
-	int name_len=strlen(name);
+	int name_len=(int)strlen(name);
 	int i;
 	for(i=name_len-1;i>=0;i--)
 	{
@@ -78,7 +78,7 @@ std::string PathUtil::toStdPath(const char* name)
 {
 	std::string ret(name);
 
-	int length=ret.size();
+	int length=(int)ret.size();
 	for(int i=0;i<length;i++)
 	{
 		if(ret[i]=='\\')
